+++ src/main/scala/io/bartholomews/fsclient/entities/oauth/v1/OAuthV1AuthorizationFramework.scala
@@ -1,7 +1,8 @@
 package io.bartholomews.fsclient.entities.oauth.v1
 
+import io.bartholomews.fsclient.codecs.StringPipe
 import io.bartholomews.fsclient.entities.oauth.AccessTokenCredentials
-import io.bartholomews.fsclient.requests.FsAuthRequest
+import io.bartholomews.fsclient.requests.FsAuthPlainText
 import org.http4s.Uri
 
 object OAuthV1AuthorizationFramework {
@@ -13,7 +14,8 @@ object OAuthV1AuthorizationFramework {
   }
 
   // https://tools.ietf.org/html/rfc5849#section-2.2
-  case class AccessTokenRequest(uri: Uri) extends FsAuthRequest.Post[Nothing, String, AccessTokenCredentials] {
-    final override val body: Option[Nothing] = None
+  case class AccessTokenRequest(uri: Uri)(implicit decoder: StringPipe[AccessTokenCredentials])
+      extends FsAuthPlainText.PostEmpty[AccessTokenCredentials] {
+    override def resDecoder: StringPipe[AccessTokenCredentials] = decoder
   }
 }