+++ src/main/scala/io/bartholomews/fsclient/implicits/package.scala
@@ -1,5 +1,0 @@
-package io.bartholomews.fsclient
-
-import io.bartholomews.fsclient.codecs.CodecSyntax
-
-package object implicits extends CodecSyntax