+++ src/main/scala/io/bartholomews/fsclient/utils/FsLogger.scala
@@ -1,6 +1,5 @@
 package io.bartholomews.fsclient.utils
 
-import cats.effect.Effect
 import cats.implicits._
 import fs2.Pipe
 import io.bartholomews.fsclient.utils.HttpTypes.ErrorOr
@@ -17,7 +16,7 @@ object FsLogger {
 
   logger.info(s"$logger started.")
 
-  private[fsclient] def logRequest[F[_]: Effect, B](request: Request[F])(maybeBody: Option[B]): Request[F] = {
+  private[fsclient] def logRequest[F[_], B](request: Request[F])(maybeBody: Option[B]): Request[F] = {
     logger.info(s"Request: ${request.method.name} [${request.uri}]")
     maybeBody.foreach(body => logger.debug(s"Request body - {\n\t$body\n}"))
     // TODO: should probably obfuscate log entries like `Authorization: Bearer <token>`
@@ -25,7 +24,7 @@ object FsLogger {
     request
   }
 
-  private[fsclient] def responseHeadersLogPipe[F[_]: Effect, T]: Pipe[F, Response[F], Response[F]] =
+  private[fsclient] def responseHeadersLogPipe[F[_], T]: Pipe[F, Response[F], Response[F]] =
     _.map { res =>
       logger.debug(s"Response status: [${res.status}]")
       // TODO: should probably obfuscate log entries like `content-security-policy`
@@ -33,19 +32,19 @@ object FsLogger {
       res
     }
 
-  private[fsclient] def rawJsonResponseLogPipe[F[_]: Effect, A]: Pipe[F, Json, Json] =
+  private[fsclient] def rawJsonResponseLogPipe[F[_], A]: Pipe[F, Json, Json] =
     _.map { json =>
       logger.debug(s"Json response - ${json.spaces2}")
       json
     }
 
-  private[fsclient] def rawPlainTextResponseLogPipe[F[_]: Effect, A]: Pipe[F, String, String] =
+  private[fsclient] def rawPlainTextResponseLogPipe[F[_], A]: Pipe[F, String, String] =
     _.map { text =>
       logger.debug(s"PlainText response - [\n\t$text\n]")
       text
     }
 
-  private[fsclient] def responseLogPipe[F[_]: Effect, A]: Pipe[F, ErrorOr[A], ErrorOr[A]] =
+  private[fsclient] def responseLogPipe[F[_], A]: Pipe[F, ErrorOr[A], ErrorOr[A]] =
     _.map(_.map { entity =>
       logger.debug(s"Response entity - {\n\t$entity\n}")
       // TODO: figure out logger shape (entity, sumologic etc)
@@ -53,7 +52,7 @@ object FsLogger {
       entity
     })
 
-  private[fsclient] def errorLogPipe[F[_]: Effect, A]: Pipe[F, Either[Throwable, A], Either[Throwable, A]] =
+  private[fsclient] def errorLogPipe[F[_], A]: Pipe[F, Either[Throwable, A], Either[Throwable, A]] =
     _.map(_.leftMap { throwable =>
       logger.error("There was a problem with the request", throwable)
       throwable